+++ plugins/boteco-plugin-subscription/src/main/java/tools/devnull/boteco/plugins/subscription/SubscriptionParameters.java
@@ -37,8 +37,8 @@ public class SubscriptionParameters {
   private final boolean requestConfirmation;
 
   public SubscriptionParameters(IncomeMessage message, String[] events) {
-    if (message.user() != null && !message.isGroup()) {
-      this.target = message.user().id();
+    if (message.user().isPresent() && !message.isGroup()) {
+      this.target = message.user().value().id();
       this.channel = "user";
     } else {
       this.target = message.isGroup() ? message.target() : message.sender().id();