+++ plugins/boteco-plugin-user/src/main/java/tools/devnull/boteco/plugins/user/spi/UserRepository.java
@@ -66,4 +66,11 @@ public interface UserRepository {
    */
   void update(User user);
 
+  /**
+   * Removes the given user from the repository
+   *
+   * @param user the user to delete
+   */
+  void delete(User user);
+
 }