+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.11.0
+
+- New Time Bomb plugin
+- Provider API improvements
+- Removed status.redhat rss parser
+- Improvements on Message interfaces
+
 ## v0.10.2 (2017-09-27)
 
 - Fixed help message for irc plugin